refactor: don't reuse transports for users in the call
Earlier we used to re-use the transports (for audio and video) for users in a call, cuz creating a new transport on demand is expensive and creating a bunch of consumers for a subsequent producer is exponentially expensive as the meet members.
But I had to refactor it cuz of one stupid reason. Chrome and Safari doesn't like the track being disabled and want it to be stopped or they'll keeps the cam on led on even if the cam is indeed off for the app :/
And people were complaining and rejoining the meet (essentially not reusing the transports XD) so might as well just do it within the meet

Let's see how expensive it will be for the CPU